# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Ai-lifeplanner API",
        "title": "Fiber API AudioSum",
        "contact": {},
        "version": "1.0"
    },
    "host": "humorous-colt-vital.ngrok-free.app",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/mood": {
            "get": {
                "description": "Get all MoodData",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Get all MoodData",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/users/mood/{id}": {
            "get": {
                "description": "Get MoodData by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Get MoodData by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            },
            "post": {
                "description": "Get MoodData by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mood"
                ],
                "summary": "Get MoodData by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Mood Data",
                        "name": "bodyMood",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.MoodResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/users/schedule": {
            "get": {
                "description": "Get all schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get all schedule",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/users/schedule/{id}": {
            "get": {
                "description": "Get schedule by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedule by  ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scheadule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Create a new schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Shcedule Data",
                        "name": "bodyUserProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.ScheduleResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/user/schedule/{id}": {
            "get": {
                "description": "Get schedule by User ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedule by User ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/entities.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.MoodResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "mood": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "entities.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "entities.ScheduleResponse": {
            "type": "object",
            "properties": {
                "available_time": {
                    "type": "string"
                },
                "busy_days": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "preferred_times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "work_hours": {
                    "type": "string"
                }
            }
        },
...
    }
}